requirements.txt
pycaret
flask
pandas
numpy
google-generativeai
dotenv